fix(state): split handler.rs tests out + fix incidental review-repro dep (CI red)
CI on PR #2529 failed on all Check/Coverage jobs. Root cause: the GREEN
commit's additions pushed src/tasks/handler.rs to 2742 LOC, tripping the
`production_src_files_under_anti_monolith_ceiling` invariant (max 2500).
This test only runs under `tests/`, which my local `cargo test --bin
agend-terminal <filter>` runs never exercised — should have run the
integration test suite too.

Fix: re-home the inline `#[cfg(test)] mod tests { ... }` block (~1040
lines) into a sibling file `src/tasks/handler/tests.rs` (the `handler/`
directory already exists — `review_repro_tasks.rs` lives there via the
same file-module pattern). handler.rs is now 1702 LOC; zero behavior
change, pure move + dedent.

The split surfaced a second, pre-existing issue: `tests/review_tasks_6.rs`'s
`handle_update_has_inlock_actor_drift_guard_tasks` static-text gate reads
the whole of handler.rs and greps for the literal phrase "attribution
would be stale" to verify the #231 owner-drift guard's error message is
wired. That phrase was NEVER actually contiguous in the production
string — it's written with a `\`-continuation across two lines for
width — so the gate was only ever passing by accident, matching a
duplicate un-wrapped copy of the same phrase inside an embedded unit
test's `assert!` (now moved to the sibling file, since it was inside the
`mod tests` block being relocated). Fixed by de-wrapping the production
error message onto one line so the gate verifies the actual production
text, per its own stated intent — not a workaround, the correct fix now
that the incidental duplication is gone.

Verified:
  cargo test --test src_file_size_invariant   -> pass
  cargo test --test review_tasks_6             -> pass
  cargo test --bin agend-terminal 2249         -> 6/6 pass
  cargo test --bin agend-terminal tasks::      -> 178/178 pass
  cargo test --tests (all other integration binaries) -> pass; the
    handful of unrelated failures seen locally (admin/api/binding/
    bootstrap/worktree_pool etc.) are this sandbox's known real-git-
    worktree concurrency contention under full-suite parallelism —
    confirmed pre-existing and unrelated to this branch in the prior
    two commits' verification notes, reproducible identically on a
    clean origin/main checkout run the same way.
  cargo clippy --bin agend-terminal -- -D warnings -> clean
  cargo fmt --check -> clean

Refs #2524 P2a-r1 (t-20260702033209237316-14977-17), PR #2529 CI-red
on commit 1744d868.

Agend-Agent: gapfix-dev
Agend-Branch: feat/2524-p2a-plan-ack-gate
Agend-Issued-At: 2026-07-02T03:33:01.802808+00:00
